use template to generalize threadPool
