+++ src/core/cuda_utils.h
@@ -27,6 +27,7 @@
 
 #include <set>
 #include "src/core/status.h"
+#include <future>
 
 #ifdef TRITON_ENABLE_GPU
 #include <cuda_runtime_api.h>
@@ -92,5 +93,37 @@ Status CheckGPUCompatibility(
 Status GetSupportedGPUs(
     std::set<int>* supported_gpus, const double min_compute_capability);
 #endif
+
+/// A struct that stores the parameters for the CopyBuffer operation.
+struct CopyBufferData {
+  CopyBufferData(
+      const std::string& msg, const TRITONSERVER_MemoryType src_memory_type,
+      const int64_t src_memory_type_id,
+      const TRITONSERVER_MemoryType dst_memory_type,
+      const int64_t dst_memory_type_id, const size_t byte_size, const void* src,
+      void* dst, cudaStream_t cuda_stream, bool* cuda_used)
+      : msg_(msg), src_memory_type_(src_memory_type),
+        src_memory_type_id_(src_memory_type_id),
+        dst_memory_type_(dst_memory_type),
+        dst_memory_type_id_(dst_memory_type_id), byte_size_(byte_size),
+        src_(src), dst_(dst), cuda_stream_(cuda_stream), cuda_used_(cuda_used)
+  {
+  }
+
+  const std::string& msg_;
+  const TRITONSERVER_MemoryType src_memory_type_;
+  const int64_t src_memory_type_id_;
+  const TRITONSERVER_MemoryType dst_memory_type_;
+  const int64_t dst_memory_type_id_;
+  const size_t byte_size_;
+  const void* src_;
+  void* dst_;
+  cudaStream_t cuda_stream_;
+  bool* cuda_used_;
+  std::promise<Status> status_;
+};
+
+// Helper around CopyBuffer
+void CopyBufferHandler(CopyBufferData* data);
 
 }}  // namespace nvidia::inferenceserver
