+++ src/core/thread_pool.h
@@ -32,7 +32,6 @@
 #include <thread>
 #include <vector>
 
-#include "src/core/cuda_utils.h"
 #include "src/core/status.h"
 #include "src/core/sync_queue.h"
 
@@ -50,12 +49,7 @@ class ThreadPool {
   ~ThreadPool();
 
   // Add <TaskData> to queue
-  Status AddTask(
-      const std::string& msg, const TRITONSERVER_MemoryType src_memory_type,
-      const int64_t src_memory_type_id,
-      const TRITONSERVER_MemoryType dst_memory_type,
-      const int64_t dst_memory_type_id, const size_t byte_size, const void* src,
-      void* dst, cudaStream_t cuda_stream, bool* cuda_used);
+  Status AddTask(std::unique_ptr<TaskData> task_data);
 
   // Run <TaskFunction> on the worker threads. Must add <TaskData> to queue using
   // AddTask before calling CompleteQueue.
@@ -69,42 +63,10 @@ class ThreadPool {
   // wait for them to finish and start from the beginning.
   Status GetNextAvailableId(int* worker_id, bool await_available);
 
-  /// A struct that stores the parameters for the CopyBuffer operation.
-  struct CopyBufferData {
-    CopyBufferData(
-        const std::string& msg, const TRITONSERVER_MemoryType src_memory_type,
-        const int64_t src_memory_type_id,
-        const TRITONSERVER_MemoryType dst_memory_type,
-        const int64_t dst_memory_type_id, const size_t byte_size,
-        const void* src, void* dst, cudaStream_t cuda_stream, bool* cuda_used)
-        : msg_(msg), src_memory_type_(src_memory_type),
-          src_memory_type_id_(src_memory_type_id),
-          dst_memory_type_(dst_memory_type),
-          dst_memory_type_id_(dst_memory_type_id), byte_size_(byte_size),
-          src_(src), dst_(dst), cuda_stream_(cuda_stream), cuda_used_(cuda_used)
-    {
-    }
-
-    const std::string& msg_;
-    const TRITONSERVER_MemoryType src_memory_type_;
-    const int64_t src_memory_type_id_;
-    const TRITONSERVER_MemoryType dst_memory_type_;
-    const int64_t dst_memory_type_id_;
-    const size_t byte_size_;
-    const void* src_;
-    void* dst_;
-    cudaStream_t cuda_stream_;
-    bool* cuda_used_;
-    std::promise<Status> status_;
-  };
-
-  // Helper around CopyBuffer
-  static void CopyBufferHandler(CopyBufferData* data);
-
   int thread_count_;
   std::vector<std::thread> worker_threads_;
   std::vector<std::future<Status>> futures_;
-  SyncQueue<std::unique_ptr<CopyBufferData>> queue_;
+  SyncQueue<std::unique_ptr<TaskData>> queue_;
 };
 
 }}  // namespace nvidia::inferenceserver
